requirements.txt
fastapi
openai
python-dotenv
pydantic==1.*
langchain
faiss-cpu
tiktoken
streamlit